Move the group type
